+++ src/components/Maps.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import {View} from 'react-native';
+
+function Maps (props){
+    return(
+        <MapView initialRegion={props.loc}>
+            <MapView.Marker
+             
+            />
+        </MapView>
+    )
+}
+
+export default Maps;
